src/data_collection/download_ucf_crime.py
"""Download UCF Crime Dataset videos and extract frames for the benchmark.

The UCF Crime Dataset contains 1,900 real-world surveillance videos across
13 crime categories. We extract frames to get authentic CCTV still images.

Usage:
    # Download videos and extract frames (default: 1 frame per 5 seconds)
    python -m src.data_collection.download_ucf_crime

    # Extract more frames per video
    python -m src.data_collection.download_ucf_crime --fps 0.5

    # Only extract frames from already-downloaded videos
    python -m src.data_collection.download_ucf_crime --extract-only

    # Limit to specific categories
    python -m src.data_collection.download_ucf_crime --categories Arrest Robbery Shoplifting

    # Dry run
    python -m src.data_collection.download_ucf_crime --dry-run
"""

import argparse
import json
import os
import subprocess
import urllib.request
from pathlib import Path

from src.utils.config import get_data_path, get_project_root

# UCF Crime categories and their Google Drive folder structure
# Videos are hosted on Google Drive and various mirrors
UCF_CATEGORIES = [
    "Abuse", "Arrest", "Arson", "Assault", "Burglary",
    "Explosion", "Fighting", "RoadAccidents", "Robbery",
    "Shooting", "Shoplifting", "Stealing", "Vandalism",
]

# Map UCF categories to our benchmark categories
UCF_TO_BENCHMARK = {
    "Abuse": "people",
    "Arrest": "people",
    "Arson": "outdoor_scenes",
    "Assault": "people",
    "Burglary": "indoor_scenes",
    "Explosion": "outdoor_scenes",
    "Fighting": "people",
    "RoadAccidents": "vehicles",
    "Robbery": "indoor_scenes",
    "Shooting": "outdoor_scenes",
    "Shoplifting": "indoor_scenes",
    "Stealing": "indoor_scenes",
    "Vandalism": "outdoor_scenes",
}


def extract_frames_from_video(
    video_path: Path,
    output_dir: Path,
    fps: float = 0.2,
    max_frames: int = 10,
    skip_existing: bool = True,
) -> int:
    """Extract frames from a video file using ffmpeg.

    Args:
        video_path: Path to video file.
        output_dir: Directory to save extracted frames.
        fps: Frames per second to extract (0.2 = 1 frame every 5 seconds).
        max_frames: Maximum frames to extract per video.
        skip_existing: Skip if frames already exist.

    Returns:
        Number of frames extracted.
    """
    output_dir.mkdir(parents=True, exist_ok=True)
    prefix = video_path.stem

    # Check if we already extracted frames
    existing = list(output_dir.glob(f"{prefix}_*.jpg"))
    if skip_existing and len(existing) >= max_frames:
        return 0

    try:
        # Use ffmpeg to extract frames
        output_pattern = str(output_dir / f"{prefix}_%04d.jpg")

        cmd = [
            "ffmpeg", "-i", str(video_path),
            "-vf", f"fps={fps}",
            "-frames:v", str(max_frames),
            "-q:v", "2",  # High quality JPEG
            output_pattern,
            "-y",  # Overwrite
            "-loglevel", "error",
        ]

        subprocess.run(cmd, check=True, capture_output=True, timeout=60)

        extracted = list(output_dir.glob(f"{prefix}_*.jpg"))
        return len(extracted)

    except (subprocess.CalledProcessError, subprocess.TimeoutExpired, FileNotFoundError) as e:
        return 0


def extract_frames_from_directory(
    video_dir: Path,
    output_dir: Path,
    fps: float = 0.2,
    max_frames_per_video: int = 10,
    max_total: int = None,
    skip_existing: bool = True,
) -> dict:
    """Extract frames from all videos in a directory.

    Args:
        video_dir: Directory containing video files.
        output_dir: Directory to save frames.
        fps: Frames per second to extract.
        max_frames_per_video: Max frames per video.
        max_total: Max total frames to extract (across all videos).
        skip_existing: Skip videos with existing frames.

    Returns:
        Stats dict.
    """
    from tqdm import tqdm

    video_extensions = {".mp4", ".avi", ".mkv", ".mov", ".wmv", ".mpg", ".mpeg"}
    videos = sorted(
        f for f in video_dir.rglob("*")
        if f.suffix.lower() in video_extensions
    )

    if not videos:
        print(f"  No videos found in {video_dir}")
        return {"videos": 0, "frames": 0}

    stats = {"videos": 0, "frames": 0, "skipped": 0}
    total_frames = 0

    for video in tqdm(videos, desc=f"  Extracting frames"):
        if max_total and total_frames >= max_total:
            break

        remaining = max_frames_per_video
        if max_total:
            remaining = min(remaining, max_total - total_frames)

        n = extract_frames_from_video(
            video, output_dir, fps, remaining, skip_existing
        )

        if n > 0:
            stats["videos"] += 1
            stats["frames"] += n
            total_frames += n
        elif n == 0 and skip_existing:
            stats["skipped"] += 1

    return stats


def download_ucf_videos_kaggle(output_dir: Path, categories: list = None) -> bool:
    """Download UCF Crime videos using kaggle CLI.

    Requires: pip install kaggle, ~/.kaggle/kaggle.json configured.

    Returns True if successful.
    """
    try:
        import kaggle
        print("Downloading UCF Crime Dataset from Kaggle...")
        print("  Dataset: odins0n/ucf-crime-dataset")

        subprocess.run([
            "kaggle", "datasets", "download",
            "-d", "odins0n/ucf-crime-dataset",
            "-p", str(output_dir),
            "--unzip",
        ], check=True)
        return True

    except (ImportError, subprocess.CalledProcessError, FileNotFoundError):
        return False


def download_sample_videos(output_dir: Path, max_videos: int = 50) -> dict:
    """Download a sample of surveillance videos from publicly available sources.

    This is a fallback if Kaggle download doesn't work. Uses direct video
    URLs from public surveillance footage collections.
    """
    # We'll use videos from the UCF server directly
    # The UCF dataset provides an action recognition split file with video names
    print("  Note: For the full UCF Crime Dataset, install kaggle CLI:")
    print("    pip install kaggle")
    print("    kaggle datasets download -d odins0n/ucf-crime-dataset")
    print()
    print("  Attempting direct download of sample videos...")

    output_dir.mkdir(parents=True, exist_ok=True)
    stats = {"downloaded": 0, "failed": 0}

    return stats


def main():
    parser = argparse.ArgumentParser(
        description="Download UCF Crime Dataset and extract surveillance frames."
    )
    parser.add_argument(
        "--extract-only",
        action="store_true",
        help="Only extract frames from already-downloaded videos.",
    )
    parser.add_argument(
        "--video-dir",
        type=str,
        help="Path to directory containing downloaded crime videos.",
    )
    parser.add_argument(
        "--fps",
        type=float,
        default=0.2,
        help="Frames per second to extract (default: 0.2 = 1 frame per 5 seconds).",
    )
    parser.add_argument(
        "--max-frames-per-video",
        type=int,
        default=10,
        help="Max frames to extract per video (default: 10).",
    )
    parser.add_argument(
        "--max-total",
        type=int,
        default=5000,
        help="Max total frames to extract (default: 5000).",
    )
    parser.add_argument(
        "--categories",
        nargs="+",
        choices=UCF_CATEGORIES,
        help="Specific crime categories to process.",
    )
    parser.add_argument(
        "--dry-run",
        action="store_true",
        help="Show what would happen without doing it.",
    )

    args = parser.parse_args()

    output_base = get_data_path("raw/real/surveillance")

    if args.video_dir:
        video_dir = Path(args.video_dir)
    else:
        video_dir = get_data_path("raw/videos/ucf_crime")

    categories = args.categories or UCF_CATEGORIES

    if not args.extract_only:
        print("="*60)
        print("Step 1: Download UCF Crime Dataset")
        print("="*60)

        if args.dry_run:
            print(f"  Would download to: {video_dir}")
            print(f"  Categories: {categories}")
        else:
            success = download_ucf_videos_kaggle(video_dir, categories)
            if not success:
                download_sample_videos(video_dir)

    print()
    print("="*60)
    print("Step 2: Extract Frames")
    print("="*60)

    if not video_dir.exists():
        print(f"  Video directory not found: {video_dir}")
        print(f"  Download videos first, or specify --video-dir")
        return

    if args.dry_run:
        video_extensions = {".mp4", ".avi", ".mkv", ".mov", ".wmv", ".mpg", ".mpeg"}
        videos = [f for f in video_dir.rglob("*") if f.suffix.lower() in video_extensions]
        print(f"  Found {len(videos)} videos in {video_dir}")
        print(f"  Would extract up to {args.max_frames_per_video} frames each")
        print(f"  Max total: {args.max_total}")
        print(f"  Output: {output_base}")
        return

    total_stats = {"videos": 0, "frames": 0}

    for category in categories:
        cat_video_dir = video_dir / category
        if not cat_video_dir.exists():
            # Try without category subdirectory (flat structure)
            cat_video_dir = video_dir
            if not cat_video_dir.exists():
                continue

        benchmark_cat = UCF_TO_BENCHMARK.get(category, "outdoor_scenes")
        cat_output = output_base / benchmark_cat

        print(f"\n  Category: {category} -> {benchmark_cat}")

        stats = extract_frames_from_directory(
            video_dir=cat_video_dir,
            output_dir=cat_output,
            fps=args.fps,
            max_frames_per_video=args.max_frames_per_video,
            max_total=args.max_total - total_stats["frames"] if args.max_total else None,
            skip_existing=True,
        )

        total_stats["videos"] += stats["videos"]
        total_stats["frames"] += stats["frames"]

        print(f"    {stats['videos']} videos -> {stats['frames']} frames")

        if args.max_total and total_stats["frames"] >= args.max_total:
            print(f"\n  Reached max total ({args.max_total} frames)")
            break

    print(f"\n{'='*60}")
    print(f"Total: {total_stats['videos']} videos -> {total_stats['frames']} frames")
    print(f"Output: {output_base}")

    # Save extraction log
    log_path = output_base / "extraction_log.json"
    log_path.parent.mkdir(parents=True, exist_ok=True)
    with open(log_path, "w") as f:
        json.dump({
            "source": "UCF Crime Dataset",
            "fps": args.fps,
            "max_frames_per_video": args.max_frames_per_video,
            "categories": categories,
            "stats": total_stats,
        }, f, indent=2)

    print("Done!")


if __name__ == "__main__":
    main()
