requirements.txt
# Core (requires Python 3.10+)
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Deep learning
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# Evaluation and visualization
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
imagehash>=4.3.0
pyyaml>=6.0.0

# API clients (for detection tools)
replicate>=0.22.0

# Optional: experiment tracking
# wandb>=0.16.0
